+++ psi4/src/psi4/libscf_solver/uhf.h
@@ -45,7 +45,7 @@ class UHF : public HF {
 
     double compute_initial_E() override;
 
-    void common_init();
+    void common_init() override;
     void setup_potential() override;
 
     // Guess mix performed?
